fixed atoi to max and min numbers. changed strlen type to size_t

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: htsang <htsang@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/08 15:24:30 by htsang            #+#    #+#             */
-/*   Updated: 2022/05/08 22:07:56 by htsang           ###   ########.fr       */
+/*   Updated: 2022/05/09 21:36:21 by htsang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (*str != '\0')
